+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Behavior:
 
  * `apictl traffic initialize` no longer waits for the traffic-proxy to become ready before exiting.
+ * Feature: Developer portal will show swagger documentation for up to five services (or more with appropriate license)
 
 ## 0.9.0 (2019-10-08)